refactor: 로또에서 사용하는 상수 관리 LottoConstant 생성으로 인한 수정
